Casted the return of the date function to an integer to be able to compare it with the global number

+++ src/Backend/Validation/PreorderLimiter.php
@@ -58,7 +58,7 @@ class PreorderLimiter {
 			$newDate = date('Y-m-d', $fifteenMinutesAfter);
 
 			// check if the new date is a holiday or a monday. (shop is closed on mondays).
-			if (date("w", $fifteenMinutesAfter) === self::CLOSING_SHOP_DAY) {
+			if ((int)date("w", $fifteenMinutesAfter) === self::CLOSING_SHOP_DAY) {
 				\System::log("new Date " . $newDate,__METHOD__,TL_ERROR,);
 				\System::log("new Date " . date("w", $fifteenMinutesAfter),__METHOD__,TL_ERROR,);
